backend/requirements.txt
# AgroIntel Backend Dependencies
Django>=5.0,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
django-environ>=0.11.2
Pillow>=10.0,<11.0
gunicorn>=22.0,<23.0
whitenoise>=6.6.0

# Database
psycopg2-binary>=2.9.9

# AI / ML
tensorflow>=2.15.0
torch>=2.0.0
ultralytics>=8.1.0
opencv-python>=4.9.0
scikit-learn>=1.4.0
numpy>=1.26.0
matplotlib>=3.8.0
